US_SW07_BGG model added
The Smet Wouters 2007 model with BGG type final accelerator has been added based on Villa (2014). Depending on the vintage the model might not find a mode with proper posterior Hessian using the basic mode compute 4 algorithm, then it should automatically change the mode compute method. On vintages tested the model ran with both expanding and rolling windows. (mode compute 4 for the former and 5 for the latter).